--- src/pages/classroom copy/classroom.css
@@ -1,38 +1,0 @@
-.Classroom {
-  text-align: center;
-}
-
-.Classroom-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .Classroom-logo {
-    animation: Classroom-logo-spin infinite 20s linear;
-  }
-}
-
-.Classroom-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.Classroom-link {
-  color: #61dafb;
-}
-
-@keyframes Classroom-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
