+++ src/pages/game/gamepage.tsx
@@ -17,17 +17,17 @@ const GamePage: React.FC = () => {
       <div className="col-span-1"></div>
       <div className="col-span-3">
         <Link to={"/game/cat"}>
-          <img src="http://steach.ssafy.io:8082/img-upload/display/my/hzlxtcxooigo.jpg" alt="" />  
+          <img src="https://steach.ssafy.io/img-upload/display/my/hzlxtcxooigo.jpg" alt="" />  
         </Link>
       </div>
       <div className="col-span-3">
         <Link to={"/game/man"}>
-          <img src="http://steach.ssafy.io:8082/img-upload/display/my/kfhjuhnvplman.jpg" alt="" />  
+          <img src="https://steach.ssafy.io/img-upload/display/my/kfhjuhnvplman.jpg" alt="" />  
         </Link>
       </div>
       <div className="col-span-3">
         <Link to={"/game/reva"}>
-          <img src="http://steach.ssafy.io:8082/img-upload/display/my/ifnnltrozgreva.jpg" alt="" />  
+          <img src="https://steach.ssafy.io/img-upload/display/my/ifnnltrozgreva.jpg" alt="" />  
         </Link>
       </div>
       <div className="col-span-1"></div>
